[ACTP] migrate PAR e2e tests to fake intake for RC keys
