+++ test/new-e2e/tests/privateactionrunner/private_action_runner_executor_nix_test.go
@@ -6,18 +6,12 @@
 package privateactionrunner
 
 import (
-	"crypto/ed25519"
-	"crypto/rand"
-	"crypto/x509"
-	"encoding/json"
-	"encoding/pem"
 	"fmt"
 	"strings"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 
 	"github.com/DataDog/datadog-agent/test/e2e-framework/components/datadog/agentparams"
 	scenec2 "github.com/DataDog/datadog-agent/test/e2e-framework/scenarios/aws/ec2"
@@ -34,16 +28,7 @@ const (
 
 	executorListeningLogLine = "Private action runner executor listening on"
 	executorReadyLogLine     = "Private action runner executor ready to accept actions"
-
-	// pkg/remoteconfig/state.ProductActionPlatformRunnerKeys
-	runnerKeysRCProduct = "AP_RUNNER_KEYS"
 )
-
-// mirrors pkg/privateactionrunner/types.RawKey's JSON shape
-type rawRCKey struct {
-	KeyType string `json:"keyType"`
-	Key     []byte `json:"key"`
-}
 
 type linuxPrivateActionRunnerExecutorSuite struct {
 	e2e.BaseSuite[environments.Host]
@@ -63,31 +48,13 @@ func TestLinuxPrivateActionRunnerExecutorSuite(t *testing.T) {
 	))
 }
 
-func (s *linuxPrivateActionRunnerExecutorSuite) pushFakeRunnerKeysConfig() {
-	t := s.T()
-
-	pub, _, err := ed25519.GenerateKey(rand.Reader)
-	require.NoError(t, err, "failed to generate fake runner key")
-
-	pubDER, err := x509.MarshalPKIXPublicKey(pub)
-	require.NoError(t, err, "failed to marshal fake runner public key")
-
-	pubPEM := pem.EncodeToMemory(&pem.Block{Type: "PUBLIC KEY", Bytes: pubDER})
-
-	payload, err := json.Marshal(rawRCKey{KeyType: "ED25519", Key: pubPEM})
-	require.NoError(t, err, "failed to marshal fake runner key config payload")
-
-	err = s.Env().FakeIntake.Client().RCAddConfig("", runnerKeysRCProduct, "fake-runner-key", "fake-runner-key", payload)
-	require.NoError(t, err, "failed to push fake runner key config to fakeintake")
-}
-
 // TestExecutorStartsAndListens launches the on-demand executor subcommand and
 // asserts it comes up: the process runs, the gRPC unix socket is created, and
 // the log reports the server listening and ready.
 func (s *linuxPrivateActionRunnerExecutorSuite) TestExecutorStartsAndListens() {
 	host := s.Env().RemoteHost
 
-	s.pushFakeRunnerKeysConfig()
+	PushFakeRunnerKeysConfig(s.T(), s.Env().FakeIntake.Client())
 
 	// run-executor is a foreground subcommand, not the packaged systemd service.
 	// Launch it detached as dd-agent so it can bind its socket under
